This is an initial commit of the book.
The organization, diagrams, and many other things will likely change in this
initial version of the text. This provides a starting point from which we can
discuss and edit the initial chapter, figure out more about structure and tone
(and so on).

Diagrams need to be split apart, LaTeX restructured and so on.

I "texi2pdf book.tex" to generate the content at the moment.