Chapter two, first commit.
I don't know if we'll want smaller commits, etc., but for the moment, I'm just
throwing content into the book as quickly as possible. For that reason, I'm not
worrying too much about committing one thing and then another.

This adds a second chapter. It probably needs a lot of TLC, given that we are
having the artist build their first circuit. Perhaps we can outsource more of
this material to somewhere else...

I still haven't finished the chapter, but wanted to get this committed.