Adding chapter 4.
This adds a chapter four, which is where channels come into being. Probably
need to consider adding process diagrams sooner (previous chapter?), and need a
lot of careful consideration of this chapter. It felt like a lot of
information, but then again, the goal is to sprint through things, and
reinforce them through repetition in the coming chapters.

We'll see how this goes.