Updates to first four chapters
There are too many to list. I've reworked all of these chapters. I think they
flow better, but I haven't attempted to follow a strict template. The code and
images were updated as well as part of this process.

Note that the OmniGraffle document is now a flat file vs. a package. As I
encounter the packages, I'm converting them. This causes fewer problems for VCS.