FIxed chapter four, some errors
I did a find/replace on [h] to [ht], which touched many things.

Chapter 4 is cleaned up. Significant changes frmo what was there, but at the
same time the text improved. Now may need to go back and do a followup to
introduce things that were not explained in the new version.

New diagram.