Structure thoughts
Removed 'XXX' and replaced it with 'FIXME'. When noting things that need to
change from this point forward, we should use 'FIXME'. This accounts for most
of the modifications.

Added "chapter-structure.txt", which describes the chapter structure that we
should follow as much as possible. Really, everything from chapter 2 on (I
think...) should be able to follow this structure pretty easily.

code/simple-servo.occ is the code from the first of the servo chapters.