Home for diagrams.
We have a "diagrams" folder that contains all kinds of diagrams.

The diagrams in that folder should be broken out by type (as per below), and
the files in that folder should be renamed to match the new chapter names. (Eg.
instead of ch1.graffle, it should be "blinkenlights.graffle").