making-things-move.tex now conforms with structure
Made some pretty significant changes to the flow of the chapter, but things
seem to be more readable now, which is the goal.
Also renamed servo_1.txt to reflect which chapter it's associated with. I'm
thinking the series structure of the servo chapters should be reflected in
the chapter names, but that can happen later.

General note, this is the second time these changes have been pushed to github,
hopefully the website will know who I am this time....