Conditional compilation for print/screen.
This is the beginnings of print/screen generation of the book. Granted, there
will be cases where we always want the PDF embedded, but there are times when
we don't. (For example, the Fritzing diagrams are huge.)

More work will be necessary, but this gets the minimum in place. I can now
auto-generate PDF/PNG from an OmniGraffle source, and then generate book output
for one or the other. More work needs to be done to automate the generation of
both forms (screen and print), but that can wait.