A new strategy...
This may be questionable in the grand scheme of things, but I'm going to
plough ahead anyway.

I have some problems with where the book was going. (As the guy writing
it, that's OK, I suppose.) Specifically, it was trying to do too many
things at once... introduction to electronics, introduction to
process-oriented programming, etc... and it didn't help people do
practical things quickly.

I'm going to put some parallel material into a new text, starting with a
new "main" file. It is going to follow the O'Reilly Arduino Cookbook in
a fairly slavish way, using their outline/structure to structure what I
write. At the least, this should help my students this semester, and it
might help our current users more than the book does at the moment.

If we decide we want to revise and rework the text that was used
previously, we still have it. I will, at some point, move things around
so there are separate directories for separate content. For now, it's
just a mess.